core_api/requirements.txt
httpx>=0.24
beautifulsoup4
lxml
tldextract
jinja2
requests
rich
jsonschema
pytest
weasyprint   
atheris
requests>=2.25.0
beautifulsoup4>=4.9.0
flask>=2.0.0
pdfkit>=1.0.0
wkhtmltopdf>=0.2
requests>=2.25.0
beautifulsoup4>=4.9.0
PyJWT>=2.0.0
pdfkit>=1.0.0
wkhtmltopdf>=0.2
argparse>=1.4.0
requests==2.28.1
beautifulsoup4==4.11.1
urllib3==1.26.12
pyjwt==2.6.0
Flask==2.2.2
python-dotenv==0.21.0
colorama==0.4.6